Use 'open -a' when possible
'open -a iTerm path' will open a new iTerm window in the current space, ignoring iTerm windows in other spaces (which is the behavior I prefer), iff the last iTerm window interacted with was in the current space. If an iTerm window in another space was interacted with most recently, iTerm will open a tab in that window instead. Oh well. At least it switches spaces even if Mission Control's "When switching to an application, switch to a space" checkbox is unchecked, unlike createTabWithDefaultProfile().