Delay a bit before sending AppleScript to iTerm
In the special cases of opening a new window while iTerm is already running, and opening iTerm with a shell working directory containing a backslash, we now delay a partial second before sending the AppleScript to change working directory to iTerm, in an attempt to make it work more reliably. We also delay for a second to let iTerm start up, in the case where it wasn't running and we're opening it with a shell working directory containing a backslash.